Change eventlistner to act on touchstart
Fix for iOS8 (iPhone 5/6)

In the old situation a double tap on the dropzone was required to open the file explorer on iOS8 devices (the first one was probably interpreted as an hover). 

This commit adds the touchstart, now a single tap on the dropzone in iOS8 should open the file explorer.